changing output name

+++ pipeline.py
@@ -12,7 +12,7 @@ from itertools import product
 from config import Config
 
 
-def fetch_era5(date, variables_long_name):
+def fetch_era5(date, variables_long_name, download_filename):
     c = cdsapi.Client()
 
     name = 'reanalysis-era5-pressure-levels'
@@ -30,7 +30,7 @@ def fetch_era5(date, variables_long_name):
 
     r = c.retrieve(name,
                    request,
-                   'tmp.nc')
+                   download_filename)
 
 
 @task()
@@ -73,24 +73,28 @@ def save_unique_variable_date_file(dates_vars):
                                  for var_long_name, var_short_name in Config.VARIABLES.items()
                                  if var_short_name in list(map(lambda x: x.lower(), variables))]
 
-    fetch_era5(chosen_date, variables_long_name)
+    download_filename = "tmp-{:04d}{:02d}{:02d}".format(chosen_date.year,
+                                                        chosen_date.month,
+                                                        chosen_date.day)
+    fetch_era5(chosen_date, variables_long_name, download_filename)
 
-    ds = xr.open_mfdataset('tmp.nc')
-    if 'expver' in list(ds.dims):
-        ds = ds.reduce(np.nansum, 'expver')
+    ds = xr.open_mfdataset(download_filename)
+    if all(pd.date_range(chosen_date, periods=24, freq='H') == ds.time.values):
+        if 'expver' in list(ds.dims):
+            ds = ds.reduce(np.nansum, 'expver')
 
-    for var in list(variables):
-        filename = "{:04d}{:02d}{:02d}_{}_ERA5_PL_REANALYSIS.nc".format(chosen_date.year,
-                                                                        chosen_date.month,
-                                                                        chosen_date.day,
-                                                                        var.upper())
+        for var in list(variables):
+            filename = "{:04d}{:02d}{:02d}_{}_ERA5_PL_REANALYSIS.nc".format(chosen_date.year,
+                                                                            chosen_date.month,
+                                                                            chosen_date.day,
+                                                                            var.upper())
 
-        ds[var.lower()].to_netcdf(filename)
-        print(filename)
-        fs.put(filename,
-               os.path.join(Config.BUCKET,
-                            filename))
-        os.remove(filename)
+            ds[var.lower()].to_netcdf(filename)
+            print(filename)
+            fs.put(filename,
+                   os.path.join(Config.BUCKET,
+                                filename))
+            os.remove(filename)
     os.remove('tmp.nc')
